# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mourya Green World Plant Nursery API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Admin authentication
  - name: plants
    description: Plant management
  - name: categories
    description: Category management
  - name: gallery
    description: Gallery management
  - name: inquiries
    description: Customer inquiry management
  - name: siteSettings
    description: Site settings and homepage management
  - name: dashboard
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── Auth ──────────────────────────────────────────────────────────────────
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Current admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

  /auth/change-password:
    post:
      operationId: changeAdminPassword
      tags: [auth]
      summary: Change admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Unauthorized

  # ── Dashboard ────────────────────────────────────────────────────────────
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ── Categories ────────────────────────────────────────────────────────────
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Plants ────────────────────────────────────────────────────────────────
  /plants:
    get:
      operationId: listPlants
      tags: [plants]
      summary: List plants (public)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"
    post:
      operationId: createPlant
      tags: [plants]
      summary: Create plant (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"

  /plants/featured:
    get:
      operationId: getFeaturedPlants
      tags: [plants]
      summary: Get featured plants for homepage
      responses:
        "200":
          description: Featured plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"

  /plants/{id}:
    get:
      operationId: getPlant
      tags: [plants]
      summary: Get plant by ID (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          description: Not found
    patch:
      operationId: updatePlant
      tags: [plants]
      summary: Update plant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
    delete:
      operationId: deletePlant
      tags: [plants]
      summary: Delete plant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Gallery ───────────────────────────────────────────────────────────────
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery images (public)
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: addGalleryImage
      tags: [gallery]
      summary: Add gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"

  /gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete gallery image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Inquiries ─────────────────────────────────────────────────────────────
  /inquiries:
    get:
      operationId: listInquiries
      tags: [inquiries]
      summary: List all inquiries (admin)
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
    post:
      operationId: submitInquiry
      tags: [inquiries]
      summary: Submit customer inquiry (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"

  /inquiries/{id}:
    patch:
      operationId: updateInquiry
      tags: [inquiries]
      summary: Mark inquiry read/unread (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
    delete:
      operationId: deleteInquiry
      tags: [inquiries]
      summary: Delete inquiry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Site Settings ─────────────────────────────────────────────────────────
  /site-settings:
    get:
      operationId: getSiteSettings
      tags: [siteSettings]
      summary: Get all site settings (public)
      responses:
        "200":
          description: Site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateSiteSettings
      tags: [siteSettings]
      summary: Update site settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"


components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [message, admin]
      properties:
        message:
          type: string
        admin:
          $ref: "#/components/schemas/AdminUser"

    AdminUser:
      type: object
      required: [id, email]
      properties:
        id:
          type: integer
        email:
          type: string

    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string

    # Dashboard
    DashboardStats:
      type: object
      required: [totalPlants, totalCategories, totalInquiries, unreadInquiries, totalGalleryImages, featuredPlants]
      properties:
        totalPlants:
          type: integer
        totalCategories:
          type: integer
        totalInquiries:
          type: integer
        unreadInquiries:
          type: integer
        totalGalleryImages:
          type: integer
        featuredPlants:
          type: integer

    ActivityItem:
      type: object
      required: [id, type, message, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        message:
          type: string
        createdAt:
          type: string

    # Category
    Category:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string

    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string

    # Plant
    Plant:
      type: object
      required: [id, name, categoryId, featured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        featured:
          type: boolean
        careInstructions:
          type: ["string", "null"]
        sunlight:
          type: ["string", "null"]
        water:
          type: ["string", "null"]
        createdAt:
          type: string

    PlantInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        categoryId:
          type: integer
        featured:
          type: boolean
        careInstructions:
          type: string
        sunlight:
          type: string
        water:
          type: string

    PlantUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        categoryId:
          type: integer
        featured:
          type: boolean
        careInstructions:
          type: string
        sunlight:
          type: string
        water:
          type: string

    # Gallery
    GalleryImage:
      type: object
      required: [id, imageUrl, createdAt]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        createdAt:
          type: string

    GalleryImageInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        caption:
          type: string

    # Inquiry
    Inquiry:
      type: object
      required: [id, name, email, message, read, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string

    InquiryInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string

    InquiryUpdate:
      type: object
      properties:
        read:
          type: boolean

    # Site Settings
    SiteSettings:
      type: object
      required: [id]
      properties:
        id:
          type: integer
        heroTitle:
          type: ["string", "null"]
        heroSubtitle:
          type: ["string", "null"]
        heroBannerUrl:
          type: ["string", "null"]
        aboutText:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        service1Title:
          type: ["string", "null"]
        service1Desc:
          type: ["string", "null"]
        service2Title:
          type: ["string", "null"]
        service2Desc:
          type: ["string", "null"]
        service3Title:
          type: ["string", "null"]
        service3Desc:
          type: ["string", "null"]
        service4Title:
          type: ["string", "null"]
        service4Desc:
          type: ["string", "null"]

    SiteSettingsUpdate:
      type: object
      properties:
        heroTitle:
          type: string
        heroSubtitle:
          type: string
        heroBannerUrl:
          type: string
        aboutText:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        service1Title:
          type: string
        service1Desc:
          type: string
        service2Title:
          type: string
        service2Desc:
          type: string
        service3Title:
          type: string
        service3Desc:
          type: string
        service4Title:
          type: string
        service4Desc:
          type: string

    UploadResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string