# lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mourya Green World Plant Nursery API
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryInput {
  name: string;
  email: string;
  phone?: string;
  message: string;
}